move PhysicalExpr
