+++ datafusion/physical-expr-core/src/sort_properties.rs
