+++ datafusion/physical-expr/src/lib.rs
@@ -29,9 +29,7 @@ mod physical_expr;
 pub mod planner;
 mod scalar_function;
 mod sort_expr;
-pub mod sort_properties;
 pub mod string_expressions;
-pub mod tree_node;
 pub mod udf;
 pub mod utils;
 pub mod window;
@@ -49,7 +47,7 @@ pub use equivalence::EquivalenceProperties;
 pub use partitioning::{Distribution, Partitioning};
 pub use physical_expr::{
     physical_exprs_bag_equal, physical_exprs_contains, physical_exprs_equal,
-    PhysicalExpr, PhysicalExprRef,
+    PhysicalExprRef,
 };
 pub use planner::{create_physical_expr, create_physical_exprs};
 pub use scalar_function::ScalarFunctionExpr;
@@ -60,3 +58,18 @@ pub use sort_expr::{
 pub use utils::{reverse_order_bys, split_conjunction};
 
 pub use aggregate::first_last::create_first_value_accumulator;
+
+// For backwards compatibility
+pub mod sort_properties {
+    pub use datafusion_physical_expr_core::sort_properties::{
+        ExprOrdering, SortProperties,
+    };
+}
+
+// For backwards compatibility
+pub mod tree_node {
+    pub use datafusion_physical_expr_core::tree_node::ExprContext;
+}
+
+// For backwards compatibility
+pub use datafusion_physical_expr_core::physical_expr::PhysicalExpr;
