+++ datafusion/physical-expr/src/physical_expr.rs
@@ -16,247 +16,13 @@
 // under the License.
 
 use std::any::Any;
-use std::fmt::{Debug, Display};
-use std::hash::{Hash, Hasher};
 use std::sync::Arc;
 
-use crate::sort_properties::SortProperties;
-use crate::utils::scatter;
-
-use arrow::array::BooleanArray;
-use arrow::compute::filter_record_batch;
-use arrow::datatypes::{DataType, Schema};
-use arrow::record_batch::RecordBatch;
-use datafusion_common::utils::DataPtr;
-use datafusion_common::{internal_err, not_impl_err, Result};
-use datafusion_expr::interval_arithmetic::Interval;
-use datafusion_expr::ColumnarValue;
-
+use datafusion_physical_expr_core::physical_expr::PhysicalExpr;
 use itertools::izip;
-
-/// `PhysicalExpr` evaluate DataFusion expressions such as `A + 1`, or `CAST(c1
-/// AS int)`.
-///
-/// `PhysicalExpr` are the physical counterpart to [`Expr`] used in logical
-/// planning, and can be evaluated directly on a [`RecordBatch`]. They are
-/// normally created from `Expr` by a [`PhysicalPlanner`] and can be created
-/// directly using [`create_physical_expr`].
-///
-/// A Physical expression knows its type, nullability and how to evaluate itself.
-///
-/// [`PhysicalPlanner`]: https://docs.rs/datafusion/latest/datafusion/physical_planner/trait.PhysicalPlanner.html
-/// [`create_physical_expr`]: crate::create_physical_expr
-/// [`Expr`]: datafusion_expr::Expr
-///
-/// # Example: Create `PhysicalExpr` from `Expr`
-/// ```
-/// # use arrow_schema::{DataType, Field, Schema};
-/// # use datafusion_common::DFSchema;
-/// # use datafusion_expr::{Expr, col, lit};
-/// # use datafusion_physical_expr::create_physical_expr;
-/// # use datafusion_expr::execution_props::ExecutionProps;
-/// // For a logical expression `a = 1`, we can create a physical expression
-/// let expr = col("a").eq(lit(1));
-/// // To create a PhysicalExpr we need 1. a schema
-/// let schema = Schema::new(vec![Field::new("a", DataType::Int32, true)]);
-/// let df_schema = DFSchema::try_from(schema).unwrap();
-/// // 2. ExecutionProps
-/// let props = ExecutionProps::new();
-/// // We can now create a PhysicalExpr:
-/// let physical_expr = create_physical_expr(&expr, &df_schema, &props).unwrap();
-/// ```
-///
-/// # Example: Executing a PhysicalExpr to obtain [`ColumnarValue`]
-/// ```
-/// # use std::sync::Arc;
-/// # use arrow_array::{cast::AsArray, BooleanArray, Int32Array, RecordBatch};
-/// # use arrow_schema::{DataType, Field, Schema};
-/// # use datafusion_common::{assert_batches_eq, DFSchema};
-/// # use datafusion_expr::{Expr, col, lit, ColumnarValue};
-/// # use datafusion_physical_expr::create_physical_expr;
-/// # use datafusion_expr::execution_props::ExecutionProps;
-/// # let expr = col("a").eq(lit(1));
-/// # let schema = Schema::new(vec![Field::new("a", DataType::Int32, true)]);
-/// # let df_schema = DFSchema::try_from(schema.clone()).unwrap();
-/// # let props = ExecutionProps::new();
-/// // Given a PhysicalExpr, for `a = 1` we can evaluate it against a RecordBatch like this:
-/// let physical_expr = create_physical_expr(&expr, &df_schema, &props).unwrap();
-/// // Input of [1,2,3]
-/// let input_batch = RecordBatch::try_from_iter(vec![
-///   ("a", Arc::new(Int32Array::from(vec![1, 2, 3])) as _)
-/// ]).unwrap();
-/// // The result is a ColumnarValue (either an Array or a Scalar)
-/// let result = physical_expr.evaluate(&input_batch).unwrap();
-/// // In this case, a BooleanArray with the result of the comparison
-/// let ColumnarValue::Array(arr) = result else {
-///  panic!("Expected an array")
-/// };
-/// assert_eq!(arr.as_boolean(), &BooleanArray::from(vec![true, false, false]));
-/// ```
-pub trait PhysicalExpr: Send + Sync + Display + Debug + PartialEq<dyn Any> {
-    /// Returns the physical expression as [`Any`] so that it can be
-    /// downcast to a specific implementation.
-    fn as_any(&self) -> &dyn Any;
-    /// Get the data type of this expression, given the schema of the input
-    fn data_type(&self, input_schema: &Schema) -> Result<DataType>;
-    /// Determine whether this expression is nullable, given the schema of the input
-    fn nullable(&self, input_schema: &Schema) -> Result<bool>;
-    /// Evaluate an expression against a RecordBatch
-    fn evaluate(&self, batch: &RecordBatch) -> Result<ColumnarValue>;
-    /// Evaluate an expression against a RecordBatch after first applying a
-    /// validity array
-    fn evaluate_selection(
-        &self,
-        batch: &RecordBatch,
-        selection: &BooleanArray,
-    ) -> Result<ColumnarValue> {
-        let tmp_batch = filter_record_batch(batch, selection)?;
-
-        let tmp_result = self.evaluate(&tmp_batch)?;
-
-        if batch.num_rows() == tmp_batch.num_rows() {
-            // All values from the `selection` filter are true.
-            Ok(tmp_result)
-        } else if let ColumnarValue::Array(a) = tmp_result {
-            scatter(selection, a.as_ref()).map(ColumnarValue::Array)
-        } else {
-            Ok(tmp_result)
-        }
-    }
-
-    /// Get a list of child PhysicalExpr that provide the input for this expr.
-    fn children(&self) -> Vec<Arc<dyn PhysicalExpr>>;
-
-    /// Returns a new PhysicalExpr where all children were replaced by new exprs.
-    fn with_new_children(
-        self: Arc<Self>,
-        children: Vec<Arc<dyn PhysicalExpr>>,
-    ) -> Result<Arc<dyn PhysicalExpr>>;
-
-    /// Computes the output interval for the expression, given the input
-    /// intervals.
-    ///
-    /// # Arguments
-    ///
-    /// * `children` are the intervals for the children (inputs) of this
-    /// expression.
-    ///
-    /// # Example
-    ///
-    /// If the expression is `a + b`, and the input intervals are `a: [1, 2]`
-    /// and `b: [3, 4]`, then the output interval would be `[4, 6]`.
-    fn evaluate_bounds(&self, _children: &[&Interval]) -> Result<Interval> {
-        not_impl_err!("Not implemented for {self}")
-    }
-
-    /// Updates bounds for child expressions, given a known interval for this
-    /// expression.
-    ///
-    /// This is used to propagate constraints down through an expression tree.
-    ///
-    /// # Arguments
-    ///
-    /// * `interval` is the currently known interval for this expression.
-    /// * `children` are the current intervals for the children of this expression.
-    ///
-    /// # Returns
-    ///
-    /// A `Vec` of new intervals for the children, in order.
-    ///
-    /// If constraint propagation reveals an infeasibility for any child, returns
-    /// [`None`]. If none of the children intervals change as a result of propagation,
-    /// may return an empty vector instead of cloning `children`. This is the default
-    /// (and conservative) return value.
-    ///
-    /// # Example
-    ///
-    /// If the expression is `a + b`, the current `interval` is `[4, 5]` and the
-    /// inputs `a` and `b` are respectively given as `[0, 2]` and `[-∞, 4]`, then
-    /// propagation would would return `[0, 2]` and `[2, 4]` as `b` must be at
-    /// least `2` to make the output at least `4`.
-    fn propagate_constraints(
-        &self,
-        _interval: &Interval,
-        _children: &[&Interval],
-    ) -> Result<Option<Vec<Interval>>> {
-        Ok(Some(vec![]))
-    }
-
-    /// Update the hash `state` with this expression requirements from
-    /// [`Hash`].
-    ///
-    /// This method is required to support hashing [`PhysicalExpr`]s.  To
-    /// implement it, typically the type implementing
-    /// [`PhysicalExpr`] implements [`Hash`] and
-    /// then the following boiler plate is used:
-    ///
-    /// # Example:
-    /// ```
-    /// // User defined expression that derives Hash
-    /// #[derive(Hash, Debug, PartialEq, Eq)]
-    /// struct MyExpr {
-    ///   val: u64
-    /// }
-    ///
-    /// // impl PhysicalExpr {
-    /// // ...
-    /// # impl MyExpr {
-    ///   // Boiler plate to call the derived Hash impl
-    ///   fn dyn_hash(&self, state: &mut dyn std::hash::Hasher) {
-    ///     use std::hash::Hash;
-    ///     let mut s = state;
-    ///     self.hash(&mut s);
-    ///   }
-    /// // }
-    /// # }
-    /// ```
-    /// Note: [`PhysicalExpr`] is not constrained by [`Hash`]
-    /// directly because it must remain object safe.
-    fn dyn_hash(&self, _state: &mut dyn Hasher);
-
-    /// The order information of a PhysicalExpr can be estimated from its children.
-    /// This is especially helpful for projection expressions. If we can ensure that the
-    /// order of a PhysicalExpr to project matches with the order of SortExec, we can
-    /// eliminate that SortExecs.
-    ///
-    /// By recursively calling this function, we can obtain the overall order
-    /// information of the PhysicalExpr. Since `SortOptions` cannot fully handle
-    /// the propagation of unordered columns and literals, the `SortProperties`
-    /// struct is used.
-    fn get_ordering(&self, _children: &[SortProperties]) -> SortProperties {
-        SortProperties::Unordered
-    }
-}
-
-impl Hash for dyn PhysicalExpr {
-    fn hash<H: Hasher>(&self, state: &mut H) {
-        self.dyn_hash(state);
-    }
-}
 
 /// Shared [`PhysicalExpr`].
 pub type PhysicalExprRef = Arc<dyn PhysicalExpr>;
-
-/// Returns a copy of this expr if we change any child according to the pointer comparison.
-/// The size of `children` must be equal to the size of `PhysicalExpr::children()`.
-pub fn with_new_children_if_necessary(
-    expr: Arc<dyn PhysicalExpr>,
-    children: Vec<Arc<dyn PhysicalExpr>>,
-) -> Result<Arc<dyn PhysicalExpr>> {
-    let old_children = expr.children();
-    if children.len() != old_children.len() {
-        internal_err!("PhysicalExpr: Wrong number of children")
-    } else if children.is_empty()
-        || children
-            .iter()
-            .zip(old_children.iter())
-            .any(|(c1, c2)| !Arc::data_ptr_eq(c1, c2))
-    {
-        Ok(expr.with_new_children(children)?)
-    } else {
-        Ok(expr)
-    }
-}
 
 pub fn down_cast_any_ref(any: &dyn Any) -> &dyn Any {
     if any.is::<Arc<dyn PhysicalExpr>>() {
